requirements.txt
pytest~=8.2.2
selenium~=4.33.0
python-dotenv~=1.0.1
Faker~=18.13.0